requirements.txt
alembic==0.9.9
avro-python3==1.8.2
bcrypt==3.1.4
bintrees==2.0.7
boto3==1.7.35
botocore==1.10.35
certifi==2018.4.16
cffi==1.11.5
chardet==3.0.4
click==6.7
confluent-kafka==0.11.4
docutils==0.14
fastavro==0.19.7
Flask==1.0.2
Flask-Bcrypt==0.7.1
Flask-Migrate==2.2.0
Flask-Script==2.0.6
Flask-SocketIO==3.0.1
Flask-SQLAlchemy==2.3.2
gdax==1.0.6
idna==2.7
itsdangerous==0.24
Jinja2==2.10
jmespath==0.9.3
kazoo==2.5.0
krakenex==2.1.0
Mako==1.0.7
MarkupSafe==1.0
psycopg2==2.7.5
pycparser==2.18
pykafka==2.8.0.dev2
python-dateutil==2.7.3
python-editor==1.0.3
python-engineio==2.1.1
python-socketio==1.9.0
redis==2.10.6
requests==2.13.0
s3transfer==0.1.13
six==1.10.0
SQLAlchemy==1.2.8
tabulate==0.8.2
urllib3==1.23
websocket-client==0.40.0
Werkzeug==0.14.1
